Consolidate the various Gauss-model versions
Most of these were not actively used and are remnants of past work
