+++ src/floris/simulation/wake_velocity/gauss.py
