+++ src/floris/simulation/wake_velocity/gaussian_model_base.py
